Move SyncState to resource package because it's not in the database (#1585)
