+++ app/src/main/kotlin/at/bitfire/davdroid/resource/LocalJtxCollection.kt
@@ -6,7 +6,6 @@ package at.bitfire.davdroid.resource
 
 import android.accounts.Account
 import android.content.ContentProviderClient
-import at.bitfire.davdroid.db.SyncState
 import at.bitfire.ical4android.JtxCollection
 import at.bitfire.ical4android.JtxCollectionFactory
 import at.bitfire.ical4android.JtxICalObject
