+++ app/src/main/kotlin/at/bitfire/davdroid/resource/LocalTaskList.kt
@@ -8,7 +8,6 @@ import android.accounts.Account
 import android.content.ContentProviderClient
 import android.content.ContentValues
 import androidx.core.content.contentValuesOf
-import at.bitfire.davdroid.db.SyncState
 import at.bitfire.ical4android.DmfsTaskList
 import at.bitfire.ical4android.DmfsTaskListFactory
 import at.bitfire.ical4android.TaskProvider
